:hammer: chore:update scripts
