+++ scripts/build.ts
@@ -5,7 +5,7 @@ import { Dirent } from 'node:fs';
 import { resolve } from 'node:path';
 
 import { Parser } from './Parser';
-import { LobeAgent } from './agentMetaSchema';
+import { LobeAgent } from './schema/agentMeta';
 import { agents, config, localesDir, meta, publicDir } from './const';
 import { checkDir, checkJSON, getLocaleAgentFileName } from './utils';
