+++ scripts/schema/agentMeta.ts
@@ -1,26 +1,6 @@
 import { z } from 'zod';
 
-export const lLMRoleTypeSchema = z.union([
-  z.literal('user'),
-  z.literal('system'),
-  z.literal('assistant'),
-  z.literal('function'),
-]);
-
-export const lLMMessageSchema = z.object({
-  content: z.string(),
-  role: lLMRoleTypeSchema,
-});
-
-export const lLMExampleSchema = z.array(lLMMessageSchema);
-
-export const lLMParamsSchema = z.object({
-  frequency_penalty: z.number().optional().default(0),
-  max_tokens: z.number().optional(),
-  presence_penalty: z.number().optional().default(0),
-  temperature: z.number().optional().default(0),
-  top_p: z.number().optional().default(1),
-});
+import { lLMFewShotsSchema, lLMParamsSchema } from './llm';
 
 export const metaDataSchema = z.object({
   avatar: z.string(),
@@ -36,7 +16,7 @@ const lobeAgentConfigSchema = z.object({
   enableCompressThreshold: z.boolean().optional(),
   enableHistoryCount: z.boolean().optional(),
   enableMaxTokens: z.boolean().optional(),
-  example: lLMExampleSchema.optional(),
+  fewShots: lLMFewShotsSchema.optional(),
   historyCount: z.number().optional(),
   inputTemplate: z.string().optional(),
   model: z.string().default('gpt-3.5-turbo').optional(),
